% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharterStock.R
\name{highcharterStock}
\alias{highcharterStock}
\title{Highcharts Stock Chart}
\usage{
highcharterStock(
  ticker = "AAPL",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
A simple wrapper of the highcharts stock gui
}
