% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartStock.R, R/highcharterStock.R
\name{highchartStock-shiny}
\alias{highchartStock-shiny}
\alias{highchartStockOutput}
\alias{renderhighchartStock}
\title{Shiny bindings for highchartStock}
\usage{
highchartStockOutput(outputId, width = "100\%", height = "400px")

renderhighchartStock(expr, env = parent.frame(), quoted = FALSE)

highchartStockOutput(outputId, width = "100\%", height = "400px")

renderhighchartStock(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a highchartStock}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using highchartStock within Shiny
applications and interactive Rmd documents.

Output and render functions for using highchartStock within Shiny
applications and interactive Rmd documents.
}
