% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartStock.R, R/highcharterStock.R
\name{highchartStock}
\alias{highchartStock}
\title{Highcharts Stock Chart}
\usage{
highchartStock(
  ticker = "AAPL",
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)

highchartStock(
  ticker = "AAPL",
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
A simple wrapper of the highcharts stock gui

A simple wrapper of the highcharts stock gui
}
